requirements.txt.txt
aiohttp
discord.py
google-generativeai
python-dotenv